requirements.txt
aiohttp==3.7.4.post0
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
aws-requests-auth==0.4.2
awscli==1.19.111
backcall==0.2.0
bleach==3.3.0
boto3==1.17.111
botocore==1.20.111
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
colorama==0.4.3
cryptography==3.4.7
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
docutils==0.15.2
elasticsearch==7.13.3
entrypoints==0.3
idna==2.10
importlib-metadata==3.10.0
ipykernel==5.3.4
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.17.2
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.1
mkl-service==2.4.0
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numexpr==2.7.3
numpy==1.20.2
olefile==0.46
packaging==21.0
pandas==1.2.5
pandocfilters==1.4.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pip==21.1.3
prometheus-client==0.11.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.20
Pygments==2.9.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole==5.1.0
QtPy==1.9.0
requests==2.25.1
rsa==4.7.2
s3transfer==0.4.2
Send2Trash==1.5.0
setuptools==52.0.0.post20210125
sip==4.19.13
six==1.16.0
terminado==0.9.4
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.36.2
widgetsnbextension==3.5.1
wit==6.0.0
yarl==1.6.3
zipp==3.5.0
